% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSDA-Nunes.R
\name{PLS}
\alias{PLS}
\title{Partial Least Squares Modeling}
\usage{
PLS(x, y, vl, da)
}
\arguments{
\item{x}{The matrix of independent variables, or stimuli assocated with the 
responses in y.  The matrix should have a number of rows equal to the number 
of observations, and a number of columns equal to the number of variables.}

\item{y}{The matrix of dependent variables, or responses assocaited with the 
stimuli in x.  The matrix should have a number of rowa equal to the number of 
observations, and a number of columns equal to the number of variables.}

\item{vl}{The number of latent variables to include in the model, this is 
essentially the number of times that the iterated least squares procedure is 
performed and the regression coefficients captured.}

\item{da}{a boolean indicator, which indicates whether the output variables 
in this case should be handeled as classes or as continuous variables for 
regression true means use discriminant analysis, and treat the outcomes/responses
as class variables.}
}
\value{
A list of several different values containing the results of the PLSDA
procedure.
}
\description{
This function is a wrapper that will allow for generaic modeling applying the
traditional Partial Least Squares Procedure where the response is either a 
numerical response, or an outcome class (which can have as many classes as
possible).
}
\details{

}
\examples{
Prepare the data from acetylene.mat in matlab for testing. 
x1 <- c(rep(1300,6),rep(1200,6),rep(1100,4))
x2 <- c(7.5,9,11,13.5,17,23,5.3,7.5,11,13.5,17,23,5.3,7.5,11,17); 
x3 <- c(0.012,0.012,0.0115,0.013,0.0135,0.012,0.04,0.038,0.0320,0.0260,
       0.0340,0.0410,0.0840,0.0980,0.0920,0.0860);
y <- c(49, 50.2,50.5,48.5,47.5,44.5,28,31.5,34.5,35,38,38.5,15,17,20.5,29.5)
ybin <- y < mean(y);
ythree <- acetylene <- data.frame(x1=x1,x2=x2,x3=x3,y=y,ybin=ybin);
pls_model <- PLS(acetylene[,1:3],acetylene[,5],2,T);
}
